+++ packages/E2ETest/windows/TreeDumpLibrary/ReactPackageProvider.cs
@@ -1,0 +1,16 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using Microsoft.ReactNative.Bridge;
+using Microsoft.ReactNative.Managed;
+
+namespace TreeDumpLibrary
+{
+    public sealed class ReactPackageProvider : IReactPackageProvider
+    {
+        public void CreatePackage(IReactPackageBuilder packageBuilder)
+        {
+            packageBuilder.AddViewManagers();
+        }
+    }
+}