+++ packages/react-native-win32/src/Libraries/Components/View/docs/ViewWin32.checklist.ts
@@ -1,0 +1,8 @@
+export const ViewWin32Status = {
+  android: 'Started',
+  ios: 'Started',
+  macos: 'Started',
+  windows: 'Started',
+  web: 'Started',
+  win32: 'Started',
+};