+++ vnext/Desktop.ABITests/ActivationFactory.cpp
@@ -4,7 +4,7 @@
 using namespace winrt;
 
 extern "C" {
-int32_t WINRT_CALL OS_RoGetActivationFactory(void *classId, guid const &iid, void **factory) noexcept;
+int32_t __stdcall OS_RoGetActivationFactory(void *classId, guid const &iid, void **factory) noexcept;
 }
 
 #ifdef _M_IX86
@@ -17,7 +17,7 @@ bool starts_with(std::wstring_view value, std::wstring_view match) noexcept {
   return 0 == value.compare(0, match.size(), match);
 }
 
-int32_t WINRT_CALL WINRT_RoGetActivationFactory(void *classId, guid const &iid, void **factory) noexcept {
+int32_t __stdcall WINRT_RoGetActivationFactory(void *classId, guid const &iid, void **factory) noexcept {
   *factory = nullptr;
   std::wstring_view name{WINRT_WindowsGetStringRawBuffer(classId, nullptr)};
   HMODULE library{nullptr};