+++ vnext/Microsoft.ReactNative.Cxx/ModuleRegistration.h
@@ -3,18 +3,17 @@
 
 #pragma once
 #include "winrt/Microsoft.ReactNative.Bridge.h"
-#include "winrt/Microsoft.ReactNative.h"
 
 #define INTERNAL_REACT_MODULE_3_ARGS(moduleClass, moduleName, eventEmitterName)                               \
   struct moduleClass;                                                                                         \
                                                                                                               \
   template <class TDummy>                                                                                     \
-  struct moduleClass##_ModuleRegistration final : ::Microsoft::ReactNative::ModuleRegistration {              \
+  struct moduleClass##_ModuleRegistration final : winrt::Microsoft::ReactNative::Bridge::ModuleRegistration { \
     moduleClass##_ModuleRegistration() noexcept                                                               \
-        : ::Microsoft::ReactNative::ModuleRegistration(moduleName, eventEmitterName) {}                       \
+        : winrt::Microsoft::ReactNative::Bridge::ModuleRegistration(moduleName) {}                            \
                                                                                                               \
     winrt::Microsoft::ReactNative::Bridge::ReactModuleProvider MakeModuleProvider() const noexcept override { \
-      return ::Microsoft::ReactNative::MakeModuleProvider<moduleClass>(ModuleName(), EventEmitterName());     \
+      return winrt::Microsoft::ReactNative::Bridge::MakeModuleProvider<moduleClass>();                        \
     }                                                                                                         \
                                                                                                               \
     static const moduleClass##_ModuleRegistration Registration;                                               \
@@ -22,24 +21,30 @@
                                                                                                               \
   template <class TDummy>                                                                                     \
   const moduleClass##_ModuleRegistration<TDummy> moduleClass##_ModuleRegistration<TDummy>::Registration;      \
-  template struct moduleClass##_ModuleRegistration<int>;
+  template struct moduleClass##_ModuleRegistration<int>;                                                      \
+                                                                                                              \
+  template <class TRegistry>                                                                                  \
+  void RegisterModule(TRegistry &registry, moduleClass *) noexcept {                                          \
+    registry.RegisterModule<moduleClass>(                                                                     \
+        moduleName, eventEmitterName, winrt::Microsoft::ReactNative::Bridge::ReactMemberId<__COUNTER__>{});   \
+  }
 
-#define INTERNAL_REACT_MODULE_2_ARGS(moduleClass, moduleName) INTERNAL_REACT_MODULE_3_ARGS(moduleClass, moduleName, "")
+#define INTERNAL_REACT_MODULE_2_ARGS(moduleClass, moduleName) INTERNAL_REACT_MODULE_3_ARGS(moduleClass, moduleName, L"")
 
-#define INTERNAL_REACT_MODULE_1_ARGS(moduleClass) INTERNAL_REACT_MODULE_2_ARGS(moduleClass, #moduleClass)
+#define INTERNAL_REACT_MODULE_1_ARGS(moduleClass) INTERNAL_REACT_MODULE_2_ARGS(moduleClass, L## #moduleClass)
 
 #define INTERNAL_REACT_MODULE_4TH_ARG(arg1, arg2, arg3, arg4, ...) arg4
 
 #define INTERNAL_REACT_MODULE_RECOMPOSER(argsWithParentheses) INTERNAL_REACT_MODULE_4TH_ARG argsWithParentheses
 
-#define INTERNAL_REACT_MODULE_MACRO_CHOOSER(...) \
-  INTERNAL_REACT_MODULE_RECOMPOSER(              \
+#define INTERNAL_REACT_MODULE(...)  \
+  INTERNAL_REACT_MODULE_RECOMPOSER( \
       (__VA_ARGS__, INTERNAL_REACT_MODULE_3_ARGS, INTERNAL_REACT_MODULE_2_ARGS, INTERNAL_REACT_MODULE_1_ARGS, ))
 
-namespace Microsoft::ReactNative {
+namespace winrt::Microsoft::ReactNative::Bridge {
 
 struct ModuleRegistration {
-  ModuleRegistration(const char *moduleName, const char *eventEmitterName) noexcept;
+  ModuleRegistration(const wchar_t *moduleName) noexcept;
 
   virtual winrt::Microsoft::ReactNative::Bridge::ReactModuleProvider MakeModuleProvider() const noexcept = 0;
 
@@ -51,17 +56,12 @@ struct ModuleRegistration {
     return m_next;
   }
 
-  const char *ModuleName() const noexcept {
+  const wchar_t *ModuleName() const noexcept {
     return m_moduleName;
   }
 
-  const char *EventEmitterName() const noexcept {
-    return m_eventEmitterName;
-  }
-
  private:
-  const char *m_moduleName;
-  const char *m_eventEmitterName;
+  const wchar_t *m_moduleName;
   const ModuleRegistration *m_next{nullptr};
 
   static const ModuleRegistration *s_head;
@@ -69,4 +69,4 @@ struct ModuleRegistration {
 
 void AddAttributedModules(winrt::Microsoft::ReactNative::Bridge::IReactPackageBuilder const &packageBuilder) noexcept;
 
-} // namespace Microsoft::ReactNative
+} // namespace winrt::Microsoft::ReactNative::Bridge