+++ vnext/Microsoft.ReactNative.SharedManaged/ViewManagerAttributes.cs
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-
 using Microsoft.ReactNative.Bridge;
+using System;
 
 namespace Microsoft.ReactNative.Managed
 {