+++ packages/react-native-windows-extended/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log - react-native-windows-extended
 
-This log was last generated on Fri, 06 Dec 2019 19:14:55 GMT and should not be manually modified.
+This log was last generated on Thu, 12 Dec 2019 22:18:13 GMT and should not be manually modified.
 
+## 0.60.43
+Thu, 12 Dec 2019 22:18:13 GMT
+
+### Patches
+
+- Updating react-native to version: 0.60.0-microsoft.31 (53619745+rnbot@users.noreply.github.com)
 ## 0.60.37
 Fri, 06 Dec 2019 19:14:55 GMT